+++ Runtime/Views/LinePresenter.cs
@@ -193,8 +193,7 @@ namespace Yarn.Unity
         [Group("Typewriter")]
         [ShowIf(nameof(typewriterStyle), TypewriterType.Custom)]
         [UnityEngine.Serialization.FormerlySerializedAs("CustomTypewriter")]
-        [MustNotBeNull("Attach a component that implements the " + nameof(IAsyncTypewriter) + " interface.")]
-        public UnityEngine.Object? customTypewriter;
+        public InterfaceContainer<IAsyncTypewriter>? customTypewriter;
 
         /// <summary>
         /// A list of <see cref="ActionMarkupHandler"/> objects that will be
@@ -277,7 +276,7 @@ namespace Yarn.Unity
                     break;
 
                 case TypewriterType.Custom:
-                    Typewriter = ValidateCustomTypewriter();
+                    Typewriter = customTypewriter?.Interface;
                     if (Typewriter == null)
                     {
                         Debug.LogWarning("Typewriter mode is set to custom but there is no typewriter set.");
@@ -289,44 +288,6 @@ namespace Yarn.Unity
                     }
                     break;
             }
-        }
-
-        void OnValidate()
-        {
-            var tw = ValidateCustomTypewriter();
-            if (tw == null)
-            {
-                customTypewriter = null;
-            }
-            else
-            {
-                customTypewriter = tw as Component;
-            }
-        }
-
-        private IAsyncTypewriter? ValidateCustomTypewriter()
-        {
-            if (customTypewriter is GameObject gameObject)
-            {
-                foreach (var component in gameObject.GetComponents<Component>())
-                {
-                    if (component is IAsyncTypewriter)
-                    {
-                        customTypewriter = component;
-                        return component as IAsyncTypewriter;
-                    }
-                }
-            }
-
-            if (customTypewriter is Component)
-            {
-                if (customTypewriter is IAsyncTypewriter)
-                {
-                    return customTypewriter as IAsyncTypewriter;
-                }
-            }
-
-            return null;
         }
 
         /// <summary>Presents a line using the configured text view.</summary>